Added En Passant rule Tweaked (de)selecting of pieces

+++ Chess/ChessMove.cpp
@@ -31,8 +31,8 @@ ChessPosition ChessMove::distance() const
 float ChessMove::vectorDistance() const
 {
 	return std::sqrtf(
-		std::powf(m_positionFrom.getX() - m_positionTo.getX(), 2) +
-		std::powf(m_positionFrom.getY() - m_positionTo.getY(), 2));
+		std::powf(static_cast<float>(m_positionFrom.getX() - m_positionTo.getX()), 2) +
+		std::powf(static_cast<float>(m_positionFrom.getY() - m_positionTo.getY()), 2));
 }
 
 std::ostream& operator<<(std::ostream& output, const ChessMove& cp)
